elastic-logger/requirements.txt
certifi==2023.7.22
elastic-transport==8.10.0
elasticsearch==8.10.1
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
pytest==7.4.3
python-dotenv==1.0.0
urllib3==2.0.7
setuptools
pytest-env
